Add CallCenterRow generation
Add code for generating a CallCenterRow.
This involved adding code to handle slowly changing dimensions.
Slowly changing dimensions are dimension tables that don't change at any
regular intervals, but do gradually change over time.  This code computes
keys for those tables as well as given a random number, whether to
update the value in a field or keep it the same.